Notification module (#38)
* Enhance notification feature

* Enhance small ui issue

+++ packages/firebase-functions/functions/src/triggers/onUpdateBooking.ts
@@ -6,6 +6,7 @@ import { createUniqueId } from "../utils/createUniqueId";
 import { Change } from "firebase-functions/v1";
 
 const listener = async (snapshot: Change<QueryDocumentSnapshot>) => {
+  const before = snapshot.before.data() as IBookings;
   const booking = snapshot.after.data() as IBookings;
   if (!booking) {
     return;
@@ -14,25 +15,22 @@ const listener = async (snapshot: Change<QueryDocumentSnapshot>) => {
   const notificationRef = Database.collection(Collection.Notification).doc(id);
 
   if (
-    booking.status === BOOKING_STATUS.BOOKED ||
-    booking.status === BOOKING_STATUS.REJECTED
+    before.status === BOOKING_STATUS.PENDING &&
+    (booking.status === BOOKING_STATUS.BOOKED ||
+      booking.status === BOOKING_STATUS.REJECTED)
   ) {
     const _notification: INotification = {
       id,
-      description: `You booking for ${booking.bookedFor} has been ${booking.status} by ${booking.createdBy?.name}`,
+      description: `Your booking #${booking.id} has been finalized.`,
       viewed: false,
-      type:
-        booking.status === BOOKING_STATUS.BOOKED
-          ? NOTIFICATION_TYPE.BOOKING_CONFIRMED
-          : NOTIFICATION_TYPE.BOOKING_REJECTED,
+      type: NOTIFICATION_TYPE.BOOKING,
       createdAt: +new Date(),
       createdBy: {
         name: booking.bookedToFutsal.name,
         id: booking.bookedToFutsal.id,
       },
-      createdFor: booking.bookedByUser.id,
-      bookedForTime: booking.bookedFor,
-      bookingId: booking.id,
+      createdFor: booking.bookedByUser?.id ?? "",
+      collectionId: booking.id ?? "",
     };
     try {
       await notificationRef.set(_notification);
@@ -40,24 +38,43 @@ const listener = async (snapshot: Change<QueryDocumentSnapshot>) => {
       throw new Error("Error while creating notification");
     }
   }
-  if (booking.status === BOOKING_STATUS.CANCELLED) {
+  if (
+    before.status === BOOKING_STATUS.BOOKED &&
+    booking.status === BOOKING_STATUS.CANCELLED
+  ) {
     const _notification: INotification = {
       id,
-      description: `The booking for ${booking.bookedFor} has been ${booking.status} by ${booking.createdBy?.name}.`,
+      description: `The booking has been finalized.`,
       viewed: false,
-      type: NOTIFICATION_TYPE.BOOKING_CANCELLED,
+      type: NOTIFICATION_TYPE.BOOKING,
       createdAt: +new Date(),
       createdBy: booking.createdBy,
       createdFor: booking.bookedToFutsal.id,
-      bookedForTime: booking.bookedFor,
-      bookingId: booking.id,
-      updatedAt: +new Date(),
+      collectionId: booking.id ?? "",
     };
     try {
       await notificationRef.set(_notification);
       await Database.collection(Collection.Bookings)
         .doc(booking?.id ?? "")
         .delete();
+    } catch (e) {
+      throw new Error("Error while creating notification");
+    }
+  }
+
+  if (before.hasPaid === false && booking.hasPaid === true) {
+    const _notification: INotification = {
+      id,
+      description: `You received a payment from ${booking.createdBy?.name}`,
+      viewed: false,
+      type: NOTIFICATION_TYPE.PAYMENT,
+      createdAt: +new Date(),
+      createdBy: booking.createdBy,
+      createdFor: booking.bookedToFutsal.id,
+      collectionId: booking.id ?? "",
+    };
+    try {
+      await notificationRef.set(_notification);
     } catch (e) {
       throw new Error("Error while creating notification");
     }
